Model entities can be deleted in the scene editor.

+++ src/commonMain/kotlin/baaahs/scene/OpenScene.kt
@@ -5,6 +5,7 @@ import baaahs.device.PixelArrayDevice
 import baaahs.mapper.FixtureMapping
 import baaahs.model.Model
 import baaahs.sm.webapi.Problem
+import baaahs.util.Logger
 
 class OpenScene(
     val model: Model,
@@ -19,6 +20,8 @@ class OpenScene(
         }
 
     companion object {
+        private val logger = Logger<OpenScene>()
+
         fun open(scene: Scene): OpenScene {
             val model = scene.model.open()
 
@@ -26,18 +29,25 @@ class OpenScene(
 
             val fixtures = buildMap<ControllerId, MutableList<FixtureMapping>> {
                 scene.fixtures.forEach { data ->
-                    val pixelArrayDeviceConfig = data.deviceConfig as? PixelArrayDevice.Config
-                    val fixtureMapping = FixtureMapping(
-                        data.entityId?.let { model.findEntity(it) },
-                        pixelArrayDeviceConfig?.pixelCount, // TODO kill this?
-                        null, // TODO kill this?
-                        data.deviceConfig,
-                        data.transportConfig
-                    )
+                    val entity = data.entityId?.let { model.findEntityByNameOrNull(it) }
 
-                    val controllerId = ControllerId.fromName(data.controllerId)
-                    getOrPut(controllerId) { arrayListOf() }
-                        .add(fixtureMapping)
+                    if (data.entityId != null && entity == null) {
+                        logger.warn { "No such entity \"${data.entityId} found in model, but there's a fixture mapping for it." }
+                    } else {
+                        val pixelArrayDeviceConfig = data.deviceConfig as? PixelArrayDevice.Config
+
+                        val fixtureMapping = FixtureMapping(
+                            entity,
+                            pixelArrayDeviceConfig?.pixelCount, // TODO kill this?
+                            null, // TODO kill this?
+                            data.deviceConfig,
+                            data.transportConfig
+                        )
+
+                        val controllerId = ControllerId.fromName(data.controllerId)
+                        getOrPut(controllerId) { arrayListOf() }
+                            .add(fixtureMapping)
+                    }
                 }
             }
